+++ CHANGELOG.md
@@ -7,7 +7,9 @@
 - fido-authenticator: Implement the largeBlobKey extension and the largeBlobs command ([fido-authenticator#38][])
 - Report errors when loading the configuration during initialization and disable opcard if an error occured ([#394][])
 - piv: Fix crash when changing PUK ([piv-authenticator#38][])
+- Fix LED during user presence check for NK3AM ([#93][])
 
+[#93]: https://github.com/Nitrokey/nitrokey-3-firmware/issues/93
 [#394]: https://github.com/Nitrokey/nitrokey-3-firmware/pull/394
 [fido-authenticator#38]: https://github.com/Nitrokey/fido-authenticator/issues/38
 [piv-authenticator#38]: https://github.com/Nitrokey/piv-authenticator/issues/38